Move /api files back out, rename local as localAuth
