+++ services/app-web/src/api/index.ts
@@ -1,0 +1,2 @@
+export { localGQLFetch } from './localGQLFetch'
+export { fakeAmplifyFetch } from './fakeAmplifyFetch'
