+++ services/app-web/src/localAuth/index.ts
@@ -1,6 +1,3 @@
-// api
-export { localGQLFetch } from './localGQLFetch'
-export { fakeAmplifyFetch } from './fakeAmplifyFetch'
 
 export { loginLocalUser, logoutLocalUser, getLoggedInUser } from './localAuth'
