+++ services/app-web/src/pages/Auth/Auth.test.tsx
@@ -11,7 +11,7 @@ import {
     userClickByRole,
 } from '../../testHelpers/jestHelpers'
 import { CognitoLogin } from './CognitoLogin'
-import { LocalLogin } from '../../local'
+import { LocalLogin } from '../../localAuth'
 import { fetchCurrentUserMock } from '../../testHelpers/apolloHelpers'
 /*  
 This file should only have basic user flows for auth. Form and implementation details are tested at the component level.
